+++ src/Entities/Misc/Projectile.java
@@ -9,7 +9,7 @@ public abstract class Projectile extends LiveEntity {
   protected int projectileDamage;
 
   protected Projectile(
-    int row,
+    double row,
     double col,
     double projectileSpeed,
     int projectileDamage,
